Update to Gradle 8.12
This was failing due to us implementing Capability, which is actually a
sealed interface. Oh wow, if only there was a way to denote that in Java
>_>.

This does now mean we need to give each test project a concrete group,
as capabilities constructed via strings cannot have empty
groups/versions.